user_auth_project/requirements.txt
api                           0.0.7
asgiref                       3.8.1
certifi                       2024.7.4
charset-normalizer            3.3.2
Django                        5.0.6
djangorestframework           3.15.2
djangorestframework-simplejwt 5.3.1
idna                          3.7
mysqlclient                   2.2.4
nose                          1.3.7
pip                           24.0
psycopg2-binary               2.9.9
PyJWT                         2.8.0
requests                      2.32.3
setuptools                    69.5.1
sqlparse                      0.5.0
tzdata                        2024.1
urllib3                       2.2.2
wheel                         0.43.0